RIscore can be used to score conformations. Moved RIscore to docking/scoring/functions. Resolved issue with different dimension of molecules coordinates by removing hydrogens instead of adding them.
