+++ oddt/scoring/functions/__init__.py
@@ -1,5 +1,6 @@
 from .RFScore import rfscore
 from .NNScore import nnscore
 from .PLECscore import PLECscore
+from .RIscore import ri_score
 
-__all__ = ['rfscore', 'nnscore', 'PLECscore']
+__all__ = ['rfscore', 'nnscore', 'PLECscore', 'ri_score']
